Fixed logging issue by renaming logging.py (#220)
* fixed logging module name clash

* fixed docs to reflect name change

* renamed logging in toctree

* changing something just so travis will re-run

* changed comma to semicolon so travis can restart
